blockchain: Simplify block template checking.
This simplifies the CheckConnectBlockTemplate by making use of the fact
that a template may only build off the current tip or the parent of the
current tip in order to remove the reorganization detach loop and all
code to attach nodes.

This can be done because there can never be any blocks that need to be
attached given the aforementioned constraints, and the only two
possibilities for the blocks that need to be detached are none in the
case the the template is building on the current tip, or the current tip
itself in the case the template is building on its parent.

In other words, there will always be 0 nodes to attach and either 0 or 1
node(s) to detach.  Consequently, it is not necessary to include any
attach code nor have a detach loop.