multi: Migration for utxo set semantics reversal.
This adds code to migrate the database to the new schema required by the
recent utxo set semantics reversal changes.

It involves removing the existing utxo set, spend journal, ticket
database, various indexes, resetting the best chain tip back to the
genesis block, and finally performing a full reindex.  The process can
be interrupted at any point and future invocations will resume from the
point it was interrupted.

In addition, some new infrastructure to remove the ticket database and
support for index versions is added to support the migration.