blockchain: Simplify force head reorgs.
This modifies the forceHeadReorganization function to remove checks that
are now duplicates due to the recently updated conversion of the
reorganization logic.

Also, correct the exported comment while here to explain what the
function does as opposed to referring to an unexported function that
won't show up in the generated documentation.