dcrutil: Block does not cache the header bytes
This fixes a bug in dcrutil where (*Block).BlockHeaderBytes would
incorrectly return a cached byte slice of the serialized block instead
of just the header. This would only happen after the cache had been
set, such as through a call to block.Bytes(), or construction via
NewBlockFromBytes or NewBlockFromBlockAndBytes.

The comments are also changed accordingly and wrapped after 80 columns.

Update TestBlock with a test for (*Block).BlockHeaderBytes that follows
(*Block).Bytes.