blockchain: Add tests for legacy seqlock behavior.
This adds tests to ensure the incorrect sequence lock behavior of the
current consensus rules is preserved.

In addition, in order to allow similar style tests in the future, this
adds a function named quickVoteActivationParams which provides a unique
set of chain parameters which allow the tests to more quickly activate
an agenda while still creating a fully valid test chain that consists of
full blocks.