blockchain: Add test logic to find deployments.
Several of the blockchain tests have similar code to find a specific
agenda id for a given version within a given set of chain parameters.
This is cumbersome to repeat and also requires hard coding the specific
version for an agenda, which can make it more difficult to use different
chain parameters since the version often differs between them.

This improves the tests accordingly by introducing common test
infrastructure to more easily work with deployments within the tests and
refactoring the tests to make use of them.

The following is an overview of the changes:
- Introduce a findDeployment function to search for a given agenda
  within a set of network parameters and return the resulting version
  and consensus deployment struct itself
- Add a findDeploymentChoice function to find and return a given choice
  within a consensus deployment
- Add a removeDeploymentTimeConstraints function to modify a given
  deployment such that it never expires to prevent false positives in
  the tests as time passes
- Update the feature deployment tests to make use of the aforementioned
  functions and thus no longer require the version to be specified
- Update the legacy and fixed sequence lock tests accordingly