[release-v1.4] chaincfg: Introduce agenda for fixlnseqlocks vote.
This adds a new definition for the upcoming agenda vote to enable the
corrected sequence locks behavior as defined by DCP0004.  It does not
include any code to make decisions or bump the versions.  It is only the
definition.

Also, bump the chaincfg module to v1.3.0 so the new definitions are
available to consumers.