blockchain: Introduce test closure generators.
This introduces a new function named generateChaingenHelpers in the
blockchain test code which returns a struct filled with closures that
encapsulate a test instance, a chaingen generator instance, and a
blockchain instance and is intended to be used to refactor out some of
the common code used in the tests that make use of chaingen.

Also, it makes use of the new test helper functionality introduced in Go
1.9 which allows any test failures to report the line that invoked the
closure as opposed to the line inside the closure itself which greatly
improves its usability in shared code.

This only includes the new infrastructure code and does not update any
of the tests.  That will be done in separate commits.