blockchain: Use harness in legacy seqlock tests.
This refactors the legacy sequence lock tests in blockchain to use the
recently introduced chaingen harness.

In order to simplify the process of activating an agenda which these
(and other) tests requires, a new function named
AdvanceFromSVHToActiveAgenda is also introduced on the chaingen harness
which, as its names suggests, generates and accepts enough blocks with
the appropriate vote bits set to reach one block prior to the specified
agenda becoming active.