mempool: Accept test mungers for vote tx.
This modifies CreateVote in the mempool test harness to accept munge
functions similar to how CreateSignedTx works.  This allows the test
code to more easily mutate the transactions as desired prior to them
being signed which provides greater flexibility.

While here, it also sets the block that is being voted on to the current
best chain tip of the harness.  While there is not currently anything in
the tests that relies on that behavior, it will allow future tests to
more easily create votes which involve the block being voted on.