dcrjson: add cointype to WalletInfoResult
This adds the active coin type to the dcrjson.WalletInfoResult.
Since watching-only wallets do not store the coin type keys, the
CoinType field is tagged as "cointype,omitempty" so that it will be
omitted when the default value of 0 is set, as is the case when the
wallet has an error of kind dcrwallet/errors.WatchingOnly.

The location of CoinType in the struct is chosen in part because it will
not change the size of the struct because of the preceding bools that
leave room in padding for a 32-bit value.