release: Freeze version 1 mempool module.
This serves as the final release of version 1 of the mempool module.
All future releases will be moving to version 2 of the module.

Consequently, it bumps the module version as follows:

- github.com/decred/dcrd/mempool@v1.2.0

It also removes the mempool override in the root module and updates it so building the
software will still produce binaries based on the v1 module until the v2
module is fully released.