connmngr: Add DialAddr config option
This allows package users to use a dial function that receives the full
address structure instead of just the protocol family and address as
strings.

This is required for some use cases in ln. While in btcsuite this was
implemented by changing the function signature of the Dial config
option, this is done here as an extra config option to prevent having to
perform a major version bump in the connmgr module.