txscript: Optimize typeOfScript pay-to-alt-pubkey.
This continues the process of converting the typeOfScript function to
use a combination of raw script analysis and the new tokenizer instead
of the far less efficient parsed opcodes.

In particular, it converts the detection of pay-to-alt-pubkey scripts to
use raw script analysis.

In order to accomplish this, it introduces two new functions.  The first
one is named extractPubKeyAltDetails and works with the raw script bytes
to simultaneously determine if the script is a pay-to-alt-pubkey script,
and in the case it is, extract and return the relevant details.  The
second new function is named isPubKeyAltScript and is defined in terms
of the former.

The extract function approach was chosen because it is common for
callers to want to only extract relevant details from a script if the
script is of the specific type.  Extracting those details requires
performing the exact same checks to ensure the script is of the correct
type, so it is more efficient to combine the two into one and define the
type determination in terms of the result so long as the extraction does
not require allocations.

It is important to note that this new implementation intentionally
tightens the following semantics as compared to the existing
implementation:

- The signature type must now be one of the two supported types versus
  allowing any single byte data push
- The public key must now be of the correct length for the given
  signature type versus allowing any size up to 512 bytes
- The public key for schnorr secp256k1 pubkeys must now be a compressed
  public key and adhere to the strict encoding requirements for them

The following is a before and after comparison of analyzing a large
script:

benchmark                    old ns/op    new ns/op    delta
---------------------------------------------------------------
BenchmarkIsAltPubKeyScript   143449       2.99         -100.00%

benchmark                    old allocs   new allocs   delta
---------------------------------------------------------------
BenchmarkIsAltPubKeyScript   1            0            -100.00%

benchmark                    old bytes    new bytes    delta
---------------------------------------------------------------
BenchmarkIsAltPubKeyScript   466944       0            -100.00%